+++ src/main/java/atwoz/atwoz/admin/command/domain/job/JobCommandRepository.java
@@ -1,0 +1,6 @@
+package atwoz.atwoz.admin.command.domain.job;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface JobCommandRepository extends JpaRepository<Job, Long> {
+}
