--- src/main/java/atwoz/atwoz/hobby/command/domain/HobbyRepository.java
@@ -1,10 +1,0 @@
-package atwoz.atwoz.hobby.command.domain;
-
-import java.util.List;
-import java.util.Set;
-
-public interface HobbyRepository {
-    long countAllByIdIsIn(Set<Long> ids);
-
-    List<Hobby> findByIdIn(Set<Long> ids);
-}
