--- src/main/java/atwoz/atwoz/hobby/command/infra/HobbyJpaRepository.java
@@ -1,13 +1,0 @@
-package atwoz.atwoz.hobby.command.infra;
-
-import atwoz.atwoz.hobby.command.domain.Hobby;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-import java.util.Set;
-
-public interface HobbyJpaRepository extends JpaRepository<Hobby, Long> {
-    long countAllByIdIsIn(Set<Long> ids);
-
-    List<Hobby> findByIdIn(Set<Long> ids);
-}
