--- src/main/java/atwoz/atwoz/job/command/domain/JobRepository.java
@@ -1,9 +1,0 @@
-package atwoz.atwoz.job.command.domain;
-
-import java.util.Optional;
-
-public interface JobRepository {
-    boolean existsById(Long id);
-
-    Optional<Job> findById(Long id);
-}
