--- src/main/java/atwoz/atwoz/job/command/infra/JobJpaRepository.java
@@ -1,7 +1,0 @@
-package atwoz.atwoz.job.command.infra;
-
-import atwoz.atwoz.job.command.domain.Job;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface JobJpaRepository extends JpaRepository<Job, Long> {
-}
